Create DHKeyAgreement2
Actual working implementation of diffiehelman as provided by oracle, link can be found here: http://ibm.co/1ZnqCtA
The best thing about this is that it only uses the official libraries provided by oracle. 

After reading through it, i can say I understand about 60% of it. It not only has the diffiehellman part but also the encryption using the shared secret that comes later. The code is working, copy it in eclipse and see for yourself.

I think it's safe to say that this should be the implementation to move forward on albeit with the necessary modification. 
let me know what you guys think.